% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_assign.R
\name{Assign.Cell.Type}
\alias{Assign.Cell.Type}
\title{Assign.Cell.Type}
\usage{
Assign.Cell.Type(ref.markers, exp.markers)
}
\arguments{
\item{ref.markers}{Markers | Cell Type dataframe from outside reference}

\item{exp.markers}{each sample's cluster has a set of overlapped markers with ref.markers}
}
\value{
A data frame with sample|cluster|type1|marker1|type2|marker2|type3|marker3
}
\description{
The function takes top 10 (if more than) of the exp.markers to decide the top three cell type a cluster
could belong to. use the logFC fold as the weight for the matrix, and output the cell type and corresponding marker
}
