% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{BarcodeToObject}
\alias{BarcodeToObject}
\title{BarcodeToObject}
\usage{
BarcodeToObject(seurat.ob, barcodes.filtered)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{barcodes.filtered}{The barcode of the cells of interests to be
extracted out from the object}
}
\value{
The filtered seurat object
}
\description{
Given the selected barcodes, the function will create a new a
new seurat object that contains only the cell of interest. Other fields will
remain unchanged
}
