% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{Build.ConserveMarkers.All}
\alias{Build.ConserveMarkers.All}
\title{Build.ConserveMarkers.All}
\usage{
Build.ConserveMarkers.All(seurat.ob)
}
\arguments{
\item{seurat.ob}{The seurat object with clustering information}
}
\value{
A list object with each cluster's conserved element as a marker
}
\description{
A wrapper function to simultaneously find conserve markers for
all the clusters
}
