% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Build.Cyto}
\alias{Build.Cyto}
\title{Build.Cyto}
\usage{
Build.Cyto(seurat.ob, x, y, x.thresh = 0, y.thresh = 0)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{x}{The gene name of the first cell marker}

\item{y}{The gene name of the second cell marker}

\item{x.thresh}{threshold to deem positive for the first cell marker}

\item{y.thresh}{threshold to deem positive for the second cell marker}
}
\value{
filtered seurat object with $cyto field added used for further
}
\description{
The four quadrant: x+/y+ = 1 x+/y- = 2  x-/y- = 3   x-/y+ = 4
The function will add a $cyto field into the current seurat object to
specify the quadant a cell belongs to
}
