% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_assign.R
\name{Build.Exp.Markers}
\alias{Build.Exp.Markers}
\title{Build.Exp.Markers}
\usage{
Build.Exp.Markers(markers, multiple = F)
}
\arguments{
\item{markers}{the marker table output from FindAllMarkers function}

\item{multiple}{if true, will take the list object from the output of lapply and flatten the returned markers with responsive cluster as the new column}
}
\value{
a exp_marker object
}
\description{
From the Seurat FindAllMarkers Output to a Input for Assign.Cell.Type
}
