% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_assign.R
\name{Build.Ref.Markers}
\alias{Build.Ref.Markers}
\title{Build.Ref.Markers}
\usage{
Build.Ref.Markers(path, sep = ",", splitter = ",")
}
\arguments{
\item{path}{the path to the csv file with reference marker}

\item{sep}{the delimiter of the scv file}

\item{splitter}{the seperator between markers}
}
\value{
a ref_markers object
}
\description{
takes in a .csv table and build up the ref.marker object for assignment of cell types. The list should be in the following format gene | cell.type1,cell.type2,....
}
