% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{Build.Seurat.Cluster}
\alias{Build.Seurat.Cluster}
\title{Build.Seurat.Cluster}
\usage{
Build.Seurat.Cluster(seurat.ob)
}
\arguments{
\item{seurat.ob}{The seurat object that does not have UMAP run and does not
hold identified clusters}
}
\value{
A seurat object with UMAP reduction and clusters annotated
}
\description{
A wrpper function for Seurat cluster building process
}
