% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DE.Each.Cluster}
\alias{DE.Each.Cluster}
\title{DE.Each.Cluster}
\usage{
DE.Each.Cluster(seurat.ob, pair = NA)
}
\arguments{
\item{seurat.ob}{The seurat object with clustering information}

\item{pair}{a pair of conditions to be differentially analyzed in the
corresponding "stim" field}
}
\value{
flattened list of differentially expressed genes table
}
\description{
A wrapper function to output: For each anchored cluster
output the differentially expressed gene between a pair
}
