% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{FeaturePlot.All}
\alias{FeaturePlot.All}
\title{FeaturePlot.All}
\usage{
FeaturePlot.All(seurat.ob, markers)
}
\arguments{
\item{seurat.ob}{The seurat object with clustering information}

\item{markers}{The marker list with each element as a dataframe from
FindMarkers output}
}
\value{
feature plots
}
\description{
A wrapper function to plot the top-three condident markers
from each cluster
}
