% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_assign.R
\name{Filter.Gene.Mat}
\alias{Filter.Gene.Mat}
\title{Filter.Gene.Mat}
\usage{
Filter.Gene.Mat(count.matrix, gene, threshold)
}
\arguments{
\item{count.matrix}{a matrix type that have a count for value and proper column name as cells and rowname as the gene}

\item{gene}{the gene gonna impose the threshold filter on}

\item{threshold}{the threshold to filter from count.matrix}
}
\value{
the filtered matrix
}
\description{
keep only the cells that have expression of the gene higher than the threshold, and output the new matrix object
}
