% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Filter.Gene}
\alias{Filter.Gene}
\title{Filter.Gene}
\usage{
Filter.Gene(seurat.ob, gene, threshold = 0, reverse = F)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{gene}{the gene gonna impose the threshold filter on}

\item{threshold}{the count number (exclusively greater) to define the
presence of a certain gene in the cell. Default to 0}

\item{reverse}{whether we would select the cells absent with the gene
count smaller or equal to the threshold}
}
\value{
return the remained cell's barcode
}
\description{
keep only the cells barcode that have expression of gene higher
than the threshold for now use seurat.ob->assays$RNA@data field,
but may change the assay type
}
