% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{Find.Markers.Each}
\alias{Find.Markers.Each}
\title{Find.Markers.Each}
\usage{
Find.Markers.Each(seurat.ob, min.pct = 0.1, logfc.threshold = 0.3,
  multi = NA)
}
\arguments{
\item{seurat.ob}{The seurat object with clustering information}

\item{min.pct}{as in Seurat::FindAllMarkers function}

\item{logfc.threshold}{the minimum log fold change threshold to
detect differentially expressed change}

\item{multi}{if "stim" field is not "CTRL" VS "STIM", then the multi field
would take in a vector with the name of the stim conditions}
}
\value{
a object for each stim condition, stored in a ret object list
}
\description{
A wrapper function to output: For each condition in the "stim"
field, each cluster's marker that are significantly overexpressed in a
given cluster compared to the rest of the clusters
}
