% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{FindConservedMarkers.flat}
\alias{FindConservedMarkers.flat}
\title{FindConservedMarkers.flat}
\usage{
FindConservedMarkers.flat(seurat.ob, ident.1, group, verbose)
}
\arguments{
\item{seurat.ob}{The seurat object with clustering information}
}
\value{
A datafrane with conserved element as a marker of a single cluster
}
\description{
A helper function for finding conserve markers of a single cluster
}
