% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{Flatten.Helper}
\alias{Flatten.Helper}
\title{Flatten.Helper}
\usage{
Flatten.Helper(ob)
}
\arguments{
\item{ob}{A object list of length = number of clusters, and each element is a dataframe of conserve markers}
}
\value{
A flattened conserve marker list ready for Shine.Out
}
\description{
A helper function to combine the dataframe list object into a single dataframe wiith cluster
}
