% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Plot.Cluster.Percentage}
\alias{Plot.Cluster.Percentage}
\title{Plot.Cluster.Percentage}
\usage{
Plot.Cluster.Percentage(seurat.ob, grouped = F, multi = NA)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{grouped}{a boolean to specify if the seurat object is grouped as
'CTRL' and 'STIM', default to false}

\item{multi}{a boolean to specify if the seurat object is grouped as
a specified vector of group names, default to NA}
}
\value{
return barplot with the number of cells of each cluster
within the sample/population
}
\description{
Plot.Cluster.Percentage
}
