% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Plot.Cyto.Cluster}
\alias{Plot.Cyto.Cluster}
\title{Plot.Cyto.Cluster}
\usage{
Plot.Cyto.Cluster(seurat.ob, x, y, split = F)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{x}{The gene name of the first cell marker}

\item{y}{The gene name of the second cell marker}

\item{split}{whether there are two conditions as CTRL and STIM}
}
\value{
Plot that map red dots on the original DimPlot to indicate the
presence of the two marker genes with four quadrants
}
\description{
Map back the cytometry assignment to the DimPlot of the
Seurat Objects clutsers
}
