% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Plot.Cyto.Count}
\alias{Plot.Cyto.Count}
\title{Plot.Cyto.Count}
\usage{
Plot.Cyto.Count(seurat.ob, x, y, split = F, return.stats = F)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{x}{The gene name of the first cell marker}

\item{y}{The gene name of the second cell marker}

\item{split}{whether there are two conditions as CTRL and STIM}

\item{return.stats}{whether to return the stats default to FALSE}
}
\value{
Plot that Count the number of cells in each quadrant
of the seurat.cyto object
}
\description{
plot the quadrant grapsh with
x+/y+ = 1 x+/y- = 2  x-/y- = 3   x-/y+ = 4
}
