% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Plot.FeatureScatter}
\alias{Plot.FeatureScatter}
\title{Plot.FeatureScatter}
\usage{
Plot.FeatureScatter(seurat.ob, x, y)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{x}{The gene name of the first cell marker}

\item{y}{The gene name of the second cell marker}
}
\value{
a grid-arranged FeatureScatter Seurat plot
}
\description{
Wrapper function for the Seurat.FeatureScatter for 2 different
conditions
}
