% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cytometry.R
\name{Recluster.Quadrant}
\alias{Recluster.Quadrant}
\title{Recluster.Quadrant}
\usage{
Recluster.Quadrant(seurat.ob, n.quadrant = 1)
}
\arguments{
\item{seurat.ob}{a seurat object type that have a count a count table inside
with proper column name as cells and rowname as the gene}

\item{n.quadrant}{the numbering of the selected quadrant that would like to
be reclustered}
}
\value{
a new seurat object only within a specified quadrant
}
\description{
with n.quadrant label:
x+/y+ = 1 x+/y- = 2  x-/y- = 3   x-/y+ = 4
return the subset of seurat ob cells that lies in a given quadrant
}
