% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{Shine.Out}
\alias{Shine.Out}
\title{DE.Each.Cluster}
\usage{
Shine.Out(ob, diff, markers.each, markers.conserved, multiple = F)
}
\arguments{
\item{multiple}{a logical boolean on with there are more than 2 conditions,
default to FALSE. If FALSE, the condition must be indicated as "CTRL" and
"STIM"}

\item{seurat.ob}{The seurat object with clustering information}

\item{pair}{a pair of conditions to be differentially analyzed in the
corresponding "stim" field}
}
\value{
A seurat out object that can be used for input into app.R for
ShinyApp visualization
}
\description{
A wrapper function to output: For each anchored cluster
output the differentially expressed gene between a pair
}
